refactor(auth): Integrate TokenRepository in social and token authentication flows
- Update social auth endpoints to include TokenRepository dependency
- Modify AuthService to accept TokenRepository in constructor
- Add rate limiting support for login endpoint
- Refactor token introspection to use role value
- Simplify password reset verification method
- Update dependencies and imports for token-related schemas and services
