+++ app/api/v1/auth/social.py
@@ -2,24 +2,24 @@
 
 import logging
 import secrets
-from enum import Enum
+from enum import StrEnum
 
 from fastapi import APIRouter, HTTPException, Request, Response, status
 from starlette.responses import RedirectResponse
 
 from app.core.auth import requires_admin
-from app.core.dependencies import AuditRepo, UserRepo
+from app.core.dependencies import AuditRepo, TokenRepo, UserRepo
 from app.core.errors import AuthError
 from app.core.oauth2 import AppleOAuthUserInfo, OAuthUserInfo, oauth
+from app.schemas.token import TokenResponse
 from app.services import AuthService
-from app.services.token import TokenResponse
 
 logger = logging.getLogger(__name__)
 
 router = APIRouter(prefix="/social", tags=["social"])
 
 
-class ProviderType(str, Enum):
+class ProviderType(StrEnum):
     """OAuth2 provider types.
 
     Supported providers:
@@ -63,6 +63,7 @@ async def oauth_callback_google(
     response: Response,
     user_repo: UserRepo,
     audit_repo: AuditRepo,
+    token_repo: TokenRepo,
     user_info: OAuthUserInfo,
 ) -> TokenResponse:
     """Handle Google OAuth2 callback.
@@ -72,6 +73,7 @@ async def oauth_callback_google(
         response: FastAPI response
         user_repo: User repository
         audit_repo: Audit log repository
+        token_repo: Token repository
         user_info: OAuth user info from Google
 
     Returns:
@@ -91,7 +93,7 @@ async def oauth_callback_google(
         # Clear the state from session after verification
         request.session.pop("oauth_state", None)
 
-        auth_service = AuthService(user_repo, audit_repo)
+        auth_service = AuthService(user_repo, audit_repo, token_repo)
         return await auth_service.handle_social_auth(
             request=request,
             user_info=user_info,
@@ -137,6 +139,7 @@ async def oauth_callback_apple(
     response: Response,
     user_repo: UserRepo,
     audit_repo: AuditRepo,
+    token_repo: TokenRepo,
     user_info: AppleOAuthUserInfo,
 ) -> TokenResponse:
     """Handle Apple OAuth2 callback.
@@ -146,6 +149,7 @@ async def oauth_callback_apple(
         response: FastAPI response
         user_repo: User repository
         audit_repo: Audit log repository
+        token_repo: Token repository
         user_info: OAuth user info from Apple
 
     Returns:
@@ -165,7 +169,7 @@ async def oauth_callback_apple(
         # Clear the state from session after verification
         request.session.pop("oauth_state", None)
 
-        auth_service = AuthService(user_repo, audit_repo)
+        auth_service = AuthService(user_repo, audit_repo, token_repo)
         return await auth_service.handle_social_auth(
             request=request,
             user_info=user_info,
@@ -191,12 +195,14 @@ async def list_providers() -> list[str]:
 async def get_social_stats(
     user_repo: UserRepo,
     audit_repo: AuditRepo,
+    token_repo: TokenRepo,
 ) -> dict[str, int]:
     """Get social login statistics for monitoring and analytics.
 
     Args:
         user_repo: User repository
         audit_repo: Audit log repository
+        token_repo: Token repository
 
     Returns:
         Dictionary with provider statistics
@@ -205,7 +211,7 @@ async def get_social_stats(
         HTTPException: If stats retrieval fails
     """
     try:
-        auth_service = AuthService(user_repo, audit_repo)
+        auth_service = AuthService(user_repo, audit_repo, token_repo)
         return await auth_service.get_provider_stats()
     except AuthError as e:
         raise HTTPException(
